New_Orm_废弃/requirements.txt
requests~=2.28.2
python-dotenv~=1.0.0
fastapi~=0.94.1
pydantic~=1.10.6
uvicorn~=0.21.0
APScheduler~=3.10.1
SQLAlchemy~=2.0.6
